requirements.txt
crewai==0.126.0
crewai-tools==0.46.0
google-generativeai==0.8.5
python-dotenv==1.1.0
PyYAML==6.0.2
sqlite3
